+++ llm_guard/transformers_helpers.py
@@ -101,6 +101,53 @@ def get_tokenizer_and_model_for_classification(
     return tf_tokenizer, tf_model
 
 
+def get_tokenizer_and_model_for_ner(
+    model: Model,
+    use_onnx: bool = False,
+):
+    """
+    This function loads a tokenizer and model given a model identifier and caches them.
+    Subsequent calls with the same model_identifier will return the cached tokenizer.
+
+    Args:
+        model (str): The model identifier to load the tokenizer and model for.
+        use_onnx (bool): Whether to use the ONNX version of the model. Defaults to False.
+    """
+    tf_tokenizer = get_tokenizer(model)
+    transformers = lazy_load_dep("transformers")
+
+    if use_onnx and is_onnx_supported() is False:
+        LOGGER.warning("ONNX is not supported on this machine. Using PyTorch instead of ONNX.")
+        use_onnx = False
+
+    if use_onnx is False:
+        tf_model = transformers.AutoModelForTokenClassification.from_pretrained(
+            model.path, subfolder=model.subfolder, revision=model.revision, **model.kwargs
+        )
+        LOGGER.debug("Initialized NER model", model=model, device=device())
+
+        return tf_tokenizer, tf_model
+
+    optimum_onnxruntime = lazy_load_dep(
+        "optimum.onnxruntime",
+        "optimum[onnxruntime]" if device().type != "cuda" else "optimum[onnxruntime-gpu]",
+    )
+
+    tf_model = optimum_onnxruntime.ORTModelForTokenClassification.from_pretrained(
+        model.onnx_path,
+        export=False,
+        subfolder=model.onnx_subfolder,
+        provider=("CUDAExecutionProvider" if device().type == "cuda" else "CPUExecutionProvider"),
+        revision=model.onnx_revision,
+        file_name=model.onnx_filename,
+        use_io_binding=True if device().type == "cuda" else False,
+        **model.kwargs,
+    )
+    LOGGER.debug("Initialized NER ONNX model", model=model, device=device())
+
+    return tf_tokenizer, tf_model
+
+
 ClassificationTask = Literal["text-classification", "zero-shot-classification"]